Extract duplicate expect code for `FeedImageDataLoader` into helper group
